requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.4.0
arrow==1.3.0
attrs==25.3.0
certifi==2025.6.15
charset-normalizer==3.4.2
frozenlist==1.7.0
idna==3.10
multidict==6.6.3
pillow==11.3.0
propcache==0.3.2
Pympler==1.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
qrcode==8.2
redis==6.2.0
requests==2.32.4
six==1.17.0
types-python-dateutil==2.9.0.20250516
typing_extensions==4.14.1
urllib3==2.5.0
yarl==1.20.1
